requirements.txt
fastapi[standard]
langchain
langgraph
langchain-openrouter
langchain-ollama
langchain-mongodb
langchain-text-splitters
langchain-community
PyPDF2
litellm
PyMuPDF
gdown
spacy
pytextrank
pytesseract
ipython
python-dotenv
typing-extensions
pymongo
bcrypt
pypdf
sentence-transformers